# speed up `prepare_data()` helper